feat: OceanHeader and OceanBottomNavigation adjustment
