+++ ocean-components/src/main/java/br/com/useblu/oceands/components/compose/OceanBottomNavigation.kt
@@ -8,15 +8,21 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.WindowInsetsSides
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.only
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBars
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
@@ -97,10 +103,16 @@ private fun OceanBottomNavigationPreview() {
         }
     }
 
-    OceanBottomNavigation(
-        selectedIndex = selectedIndex.value,
-        models = initialModelList.value
-    )
+    Scaffold(
+        bottomBar = {
+            OceanBottomNavigation(
+                selectedIndex = selectedIndex.value,
+                models = initialModelList.value
+            )
+        }
+    ) {
+        it
+    }
 }
 
 @Composable
@@ -123,6 +135,11 @@ fun OceanBottomNavigation(
         modifier = Modifier
             .fillMaxWidth()
             .background(color = OceanColors.brandPrimaryPure)
+            .windowInsetsPadding(
+                WindowInsets.systemBars.only(
+                    WindowInsetsSides.Bottom
+                )
+            )
             .padding(OceanSpacing.xxxs)
             .height(64.dp)
     ) {
